fixed issues with domains

+++ src/sempy_labs/admin/_domains.py
@@ -6,30 +6,78 @@ from sempy_labs.admin._basic_functions import list_workspaces
 from sempy_labs._helper_functions import (
     _base_api,
     _create_dataframe,
+    _is_valid_uuid,
 )
 
 
-def resolve_domain_id(domain_name: str) -> UUID:
+def resolve_domain_id(domain: Optional[str | UUID] = None, **kwargs) -> UUID:
     """
     Obtains the domain Id for a given domain name.
 
     Parameters
     ----------
-    domain_name : str
-        The domain name
+    domain_name : str | uuid.UUID
+        The domain name or ID
 
     Returns
     -------
-    UUID
+    uuid.UUID
         The domain Id.
     """
 
+    if "domain_name" in kwargs:
+        domain = kwargs["domain_name"]
+        print(
+            f"{icons.warning} The 'domain_name' parameter is deprecated. Please use 'domain' instead."
+        )
+
+    if domain is None:
+        raise ValueError(f"{icons.red_dot} Please provide a domain.")
+
+    if _is_valid_uuid(domain):
+        return domain
+
     dfL = list_domains()
-    dfL_filt = dfL[dfL["Domain Name"] == domain_name]
-    if len(dfL_filt) == 0:
-        raise ValueError(f"{icons.red_dot} '{domain_name}' is not a valid domain name.")
+    dfL_filt = dfL[dfL["Domain Name"] == domain]
+    if dfL_filt.empty:
+        raise ValueError(f"{icons.red_dot} '{domain}' is not a valid domain name.")
 
     return dfL_filt["Domain ID"].iloc[0]
+
+
+def resolve_domain_name(domain: Optional[str | UUID], **kwargs) -> UUID:
+    """
+    Obtains the domain name for a given domain ID.
+
+    Parameters
+    ----------
+    domain : str | uuid.UUID
+        The domain name or ID
+
+    Returns
+    -------
+    str
+        The domain Name.
+    """
+
+    if "domain_name" in kwargs:
+        domain = kwargs["domain_name"]
+        print(
+            f"{icons.warning} The 'domain_name' parameter is deprecated. Please use 'domain' instead."
+        )
+
+    if domain is None:
+        raise ValueError(f"{icons.red_dot} Please provide a domain.")
+
+    if not _is_valid_uuid(domain):
+        return domain
+
+    dfL = list_domains()
+    dfL_filt = dfL[dfL["Domain ID"] == domain]
+    if dfL_filt.empty:
+        raise ValueError(f"{icons.red_dot} '{domain}' is not a valid domain name.")
+
+    return dfL_filt["Domain Name"].iloc[0]
 
 
 def list_domains(non_empty_only: bool = False) -> pd.DataFrame:
@@ -80,7 +128,7 @@ def list_domains(non_empty_only: bool = False) -> pd.DataFrame:
     return df
 
 
-def list_domain_workspaces(domain_name: str) -> pd.DataFrame:
+def list_domain_workspaces(domain: Optional[str] = None, **kwargs) -> pd.DataFrame:
     """
     Shows a list of workspaces within the domain.
 
@@ -90,8 +138,8 @@ def list_domain_workspaces(domain_name: str) -> pd.DataFrame:
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
 
     Returns
     -------
@@ -99,7 +147,16 @@ def list_domain_workspaces(domain_name: str) -> pd.DataFrame:
         A pandas dataframe showing a list of workspaces within the domain.
     """
 
-    domain_id = resolve_domain_id(domain_name)
+    if "domain_name" in kwargs:
+        domain = kwargs["domain_name"]
+        print(
+            f"{icons.warning} The 'domain_name' parameter is deprecated. Please use 'domain' instead."
+        )
+
+    if domain is None:
+        raise ValueError(f"{icons.red_dot} Please provide a domain.")
+
+    domain_id = resolve_domain_id(domain)
 
     columns = {
         "Workspace ID": "string",
@@ -124,7 +181,8 @@ def list_domain_workspaces(domain_name: str) -> pd.DataFrame:
 def create_domain(
     domain_name: str,
     description: Optional[str] = None,
-    parent_domain_name: Optional[str] = None,
+    parent_domain: Optional[str | UUID] = None,
+    **kwargs,
 ):
     """
     Creates a new domain.
@@ -137,18 +195,24 @@ def create_domain(
         The domain name.
     description : str, default=None
         The domain description.
-    parent_domain_name : str, default=None
-        The parent domain name.
+    parent_domain : str | uuid.UUID, default=None
+        The parent domain name or ID.
     """
 
-    if parent_domain_name is not None:
-        parent_domain_id = resolve_domain_id(parent_domain_name)
+    if "parent_domain_name" in kwargs:
+        parent_domain = kwargs["parent_domain_name"]
+        print(
+            f"{icons.warning} The 'parent_domain_name' parameter is deprecated. Please use 'parent_domain' instead."
+        )
+
+    if parent_domain is not None:
+        parent_domain_id = resolve_domain_id(domain=parent_domain)
 
     payload = {}
     payload["displayName"] = domain_name
     if description is not None:
         payload["description"] = description
-    if parent_domain_name is not None:
+    if parent_domain is not None:
         payload["parentDomainId"] = parent_domain_id
 
     _base_api(
@@ -158,7 +222,7 @@ def create_domain(
     print(f"{icons.green_dot} The '{domain_name}' domain has been created.")
 
 
-def delete_domain(domain_name: str):
+def delete_domain(domain: Optional[str | UUID], **kwargs):
     """
     Deletes a domain.
 
@@ -166,20 +230,30 @@ def delete_domain(domain_name: str):
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
     """
 
-    domain_id = resolve_domain_id(domain_name)
+    if "domain_name" in kwargs:
+        domain = kwargs["domain_name"]
+        print(
+            f"{icons.warning} The 'domain_name' parameter is deprecated. Please use 'domain' instead."
+        )
+
+    if domain is None:
+        raise ValueError(f"{icons.red_dot} Please provide a domain.")
+
+    domain_id = resolve_domain_id(domain)
     _base_api(request=f"/v1/admin/domains/{domain_id}", method="delete")
 
-    print(f"{icons.green_dot} The '{domain_name}' domain has been deleted.")
+    print(f"{icons.green_dot} The '{domain}' domain has been deleted.")
 
 
 def update_domain(
-    domain_name: str,
+    domain: Optional[str | UUID] = None,
     description: Optional[str] = None,
     contributors_scope: Optional[str] = None,
+    **kwargs,
 ):
     """
     Updates a domain's properties.
@@ -188,14 +262,22 @@ def update_domain(
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
     description : str, default=None
         The domain description.
     contributors_scope : str, default=None
         The domain `contributor scope <https://learn.microsoft.com/rest/api/fabric/admin/domains/update-domain?tabs=HTTP#contributorsscopetype>`_.
     """
 
+    if "domain_name" in kwargs:
+        domain = kwargs["domain_name"]
+        print(
+            f"{icons.warning} The 'domain_name' parameter is deprecated. Please use 'domain' instead."
+        )
+
+    if domain is None:
+        raise ValueError(f"{icons.red_dot} Please provide a domain.")
     contributors_scopes = ["AdminsOnly", "AllTenant", "SpecificUsersAndGroups"]
 
     if contributors_scope not in contributors_scopes:
@@ -203,7 +285,8 @@ def update_domain(
             f"{icons.red_dot} Invalid contributors scope. Valid options: {contributors_scopes}."
         )
 
-    domain_id = resolve_domain_id(domain_name)
+    domain_id = resolve_domain_id(domain)
+    domain_name = resolve_domain_name(domain)
 
     payload = {}
     payload["displayName"] = domain_name
@@ -219,7 +302,9 @@ def update_domain(
 
 
 def assign_domain_workspaces_by_capacities(
-    domain_name: str, capacity_names: str | List[str]
+    domain: Optional[str | UUID] = None,
+    capacity_names: str | List[str],
+    **kwargs,
 ):
     """
     Assigns all workspaces that reside on the specified capacities to the specified domain.
@@ -228,15 +313,23 @@ def assign_domain_workspaces_by_capacities(
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
     capacity_names : str | List[str]
         The capacity names.
     """
 
     from sempy_labs.admin import list_capacities
 
-    domain_id = resolve_domain_id(domain_name)
+    if "domain_name" in kwargs:
+        domain = kwargs["domain_name"]
+        print(
+            f"{icons.warning} The 'domain_name' parameter is deprecated. Please use 'domain' instead."
+        )
+
+    if domain is None:
+        raise ValueError(f"{icons.red_dot} Please provide a domain.")
+    domain_id = resolve_domain_id(domain)
 
     if isinstance(capacity_names, str):
         capacity_names = [capacity_names]
@@ -272,11 +365,13 @@ def assign_domain_workspaces_by_capacities(
     )
 
     print(
-        f"{icons.green_dot} The workspaces in the {capacity_names} capacities have been assigned to the '{domain_name}' domain."
+        f"{icons.green_dot} The workspaces in the {capacity_names} capacities have been assigned to the '{domain}' domain."
     )
 
 
-def assign_domain_workspaces(domain_name: str, workspace_names: str | List[str]):
+def assign_domain_workspaces(
+    domain: Optional[str | UUID], workspace_names: str | List[str]
+):
     """
     Assigns workspaces to the specified domain by workspace.
 
@@ -284,13 +379,13 @@ def assign_domain_workspaces(domain_name: str, workspace_names: str | List[str])
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
     workspace_names : str | List[str]
         The Fabric workspace(s).
     """
 
-    domain_id = resolve_domain_id(domain_name=domain_name)
+    domain_id = resolve_domain_id(domain)
 
     if isinstance(workspace_names, str):
         workspace_names = [workspace_names]
@@ -325,11 +420,11 @@ def assign_domain_workspaces(domain_name: str, workspace_names: str | List[str])
     )
 
     print(
-        f"{icons.green_dot} The {workspace_names} workspaces have been assigned to the '{domain_name}' domain."
+        f"{icons.green_dot} The {workspace_names} workspaces have been assigned to the '{domain}' domain."
     )
 
 
-def unassign_all_domain_workspaces(domain_name: str):
+def unassign_all_domain_workspaces(domain: str | UUID):
     """
     Unassigns all workspaces from the specified domain.
 
@@ -337,11 +432,11 @@ def unassign_all_domain_workspaces(domain_name: str):
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
     """
 
-    domain_id = resolve_domain_id(domain_name=domain_name)
+    domain_id = resolve_domain_id(domain)
 
     _base_api(
         request=f"/v1/admin/domains/{domain_id}/unassignAllWorkspaces",
@@ -351,11 +446,14 @@ def unassign_all_domain_workspaces(domain_name: str):
     )
 
     print(
-        f"{icons.green_dot} All workspaces assigned to the '{domain_name}' domain have been unassigned."
+        f"{icons.green_dot} All workspaces assigned to the '{domain}' domain have been unassigned."
     )
 
 
-def unassign_domain_workspaces(domain_name: str, workspace_names: str | List[str]):
+def unassign_domain_workspaces(
+    domain: str | UUID,
+    workspace_names: str | List[str],
+):
     """
     Unassigns workspaces from the specified domain by workspace.
 
@@ -363,13 +461,13 @@ def unassign_domain_workspaces(domain_name: str, workspace_names: str | List[str
 
     Parameters
     ----------
-    domain_name : str
-        The domain name.
+    domain : str | uuid.UUID
+        The domain name or ID.
     workspace_names : str | List[str]
         The Fabric workspace(s).
     """
 
-    domain_id = resolve_domain_id(domain_name=domain_name)
+    domain_id = resolve_domain_id(domain)
 
     if isinstance(workspace_names, str):
         workspace_names = [workspace_names]
@@ -402,5 +500,5 @@ def unassign_domain_workspaces(domain_name: str, workspace_names: str | List[str
     )
 
     print(
-        f"{icons.green_dot} The {workspace_names} workspaces assigned to the '{domain_name}' domain have been unassigned."
+        f"{icons.green_dot} The {workspace_names} workspaces assigned to the '{domain}' domain have been unassigned."
     )
